+++ cmd/backup.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"path/filepath"
 	"time"
 
 	"github.com/paulschiretz/pgl-backup/pkg/buildinfo"
@@ -35,34 +34,24 @@ func RunBackup(ctx context.Context, flagMap map[string]any) error {
 
 	var err error
 
-	// Validate Base
-	base, err = util.ExpandPath(base)
+	// Validate base path
+	absBasePath, err := util.ExpandedDenormalizedAbsPath(base)
 	if err != nil {
-		return fmt.Errorf("could not expand base path: %w", err)
+		return fmt.Errorf("base path invalid: %w", err)
 	}
-	absBasePath, err := filepath.Abs(base)
-	if err != nil {
-		return fmt.Errorf("could not determine absolute base path for %s: %w", base, err)
-	}
-	absBasePath = util.DenormalizePath(absBasePath)
 
-	// NOTE: Base needs to exist, for a Backup run
+	// NOTE: Base path needs to exist, for a Backup run
 	if _, err := os.Stat(absBasePath); os.IsNotExist(err) {
 		return fmt.Errorf("base path '%s' does not exist", absBasePath)
 	}
 
-	// Validate Source
-	source, err = util.ExpandPath(source)
+	// Validate source path
+	absSourcePath, err := util.ExpandedDenormalizedAbsPath(source)
 	if err != nil {
-		return fmt.Errorf("could not expand source path: %w", err)
+		return fmt.Errorf("source path invalid: %w", err)
 	}
-	absSourcePath, err := filepath.Abs(source)
-	if err != nil {
-		return fmt.Errorf("could not determine absolute source path: %w", err)
-	}
-	absSourcePath = util.DenormalizePath(absSourcePath)
 
-	// NOTE: Source needs to exist, for a Backup run
+	// NOTE: Source path needs to exist, for a Backup run
 	if _, err := os.Stat(absSourcePath); os.IsNotExist(err) {
 		return fmt.Errorf("source path '%s' does not exist", absSourcePath)
 	}